requirements.txt
flask
torch
transformers
flask_cors
google-generativeai